backend/requirements.txt
# SHA1:6ab5a5f6b7f3f48d0e9088517171c3a0c69c981a
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
cachetools==5.5.0
    # via google-auth
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.6
    # via -r backend/requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
google-api-core==2.24.0
    # via google-api-python-client
google-api-python-client==2.156.0
    # via -r backend/requirements.in
google-auth==2.37.0
    # via
    #   -r backend/requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r backend/requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.1
    # via -r backend/requirements.in
googleapis-common-protos==1.66.0
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.5
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
oauthlib==3.2.2
    # via requests-oauthlib
proto-plus==1.25.0
    # via google-api-core
protobuf==5.29.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2-binary==2.9.10
    # via -r backend/requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.4
    # via
    #   fastapi
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.0
    # via -r backend/requirements.in
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via httplib2
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.12.0
    # via fastapi-cli
rsa==4.9
    # via google-auth
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via sqlmodel
sqlmodel==0.0.22
    # via -r backend/requirements.in
starlette==0.41.3
    # via fastapi
typer==0.15.1
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.3.0
    # via requests
uvicorn[standard]==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.3
    # via uvicorn
websockets==14.1
    # via uvicorn
